+++ anomalib/core/model/anomaly_module.py
@@ -14,7 +14,6 @@
 # See the License for the specific language governing permissions
 # and limitations under the License.
 
-from abc import ABC
 from typing import List, Union
 
 import pytorch_lightning as pl
@@ -31,7 +30,7 @@ from anomalib.core.metrics import (
 )
 
 
-class AnomalyModule(pl.LightningModule, ABC):
+class AnomalyModule(pl.LightningModule):
     """AnomalyModule to train, validate, predict and test images.
 
     Acts as a base class for all the Anomaly Modules in the library.
