+++ anomalib/data/inference.py
@@ -1,4 +1,4 @@
-"""Test Inference Dataset."""
+"""Inference Dataset."""
 
 # Copyright (C) 2020 Intel Corporation
 #
@@ -15,43 +15,13 @@
 # and limitations under the License.
 
 from pathlib import Path
-from typing import Any, List, Optional, Tuple, Union
+from typing import Any, Optional, Tuple, Union
 
 import albumentations as A
 from torch.utils.data.dataset import Dataset
-from torchvision.datasets.folder import IMG_EXTENSIONS
 
 from anomalib.data.transforms import PreProcessor
-from anomalib.data.utils import read_image
-
-
-def get_image_filenames(path: Union[str, Path]) -> List[str]:
-    """Get image filenames.
-
-    Args:
-        path (Union[str, Path]): Path to image or image-folder.
-
-    Returns:
-        List[str]: List of image filenames
-
-    """
-    image_filenames: List[str]
-
-    if isinstance(path, str):
-        path = Path(path)
-
-    # If `path` is an image path.
-    if path.is_file() and path.suffix in IMG_EXTENSIONS:
-        image_filenames = [str(path)]
-
-    # If it is a path to image folder
-    if path.is_dir():
-        image_filenames = [str(p) for p in path.glob("**/*") if p.suffix in IMG_EXTENSIONS]
-
-    if len(image_filenames) == 0:
-        raise ValueError(f"Found 0 images in {path}")
-
-    return image_filenames
+from anomalib.data.utils import get_image_filenames, read_image
 
 
 class InferenceDataset(Dataset):
@@ -60,13 +30,29 @@ class InferenceDataset(Dataset):
     def __init__(
         self,
         path: Union[str, Path],
+        pre_process: Optional[PreProcessor] = None,
         image_size: Optional[Union[int, Tuple[int, int]]] = None,
         transform_config: Optional[Union[str, A.Compose]] = None,
     ) -> None:
+        """Inference Dataset to perform prediction.
+
+        Args:
+            path (Union[str, Path]): Path to an image or image-folder.
+            pre_process (Optional[PreProcessor], optional): Pre-Processing transforms to
+                pre-process the input dataset. Defaults to None.
+            image_size (Optional[Union[int, Tuple[int, int]]], optional): Target image size
+                to resize the original image. Defaults to None.
+            transform_config (Optional[Union[str, A.Compose]], optional): Configuration file
+                parse the albumentation transforms. Defaults to None.
+        """
         super().__init__()
 
         self.image_filenames = get_image_filenames(path)
-        self.pre_process = PreProcessor(config=transform_config, image_size=image_size)
+
+        if pre_process is None:
+            self.pre_process = PreProcessor(transform_config, image_size)
+        else:
+            self.pre_process = pre_process
 
     def __len__(self) -> int:
         """Get the number of images in the given path."""
